improved services by tdd also improved the test by refactoring

+++ src/main/java/com/issuetracker/dataJpa/service/IssueServiceImpl.java
@@ -2,6 +2,8 @@ package com.issuetracker.dataJpa.service;
 
 import com.issuetracker.dataJpa.dao.IssueDao;
 import com.issuetracker.dataJpa.entity.Issue;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,7 +12,7 @@ import java.util.Optional;
 
 @Service
 public class IssueServiceImpl implements IssueService {
-
+    protected static final Logger LOGGER = LogManager.getLogger(IssueServiceImpl.class);
     private IssueDao issueDao;
     @Autowired
     public IssueServiceImpl(IssueDao issueDao) {
@@ -25,15 +27,13 @@ public class IssueServiceImpl implements IssueService {
     @Override
     public Issue findById(int theId) {
         Optional<Issue> result = issueDao.findById(theId);
-        Issue theIssue = null;
-
         if(result.isPresent()){
-            theIssue = result.get();
+            LOGGER.info("issue has been found with given id "+theId +" -> "+result.get());
+            return result.get();
         }
         else{
             throw new RuntimeException("Did not find issue with given id "+theId);
         }
-        return theIssue;
     }
 
     @Override
@@ -42,7 +42,21 @@ public class IssueServiceImpl implements IssueService {
     }
 
     @Override
-    public Issue deleteById(int theId) {
-        return issueDao.deleteById(theId);
+    public void deleteById(int theId) {
+        Optional<Issue> found = Optional.ofNullable(findById(theId));
+        if(found.isPresent()){
+            try{
+                LOGGER.info("issue is being deleted "+found.get());
+                issueDao.deleteById(theId);
+
+            }catch (Exception e){
+                LOGGER.info("Issue with given id was found but could not delete it "+found.get());
+
+            }
+
+        }else{
+            LOGGER.info("Did not find issue with given id "+theId);
+        }
+
     }
 }
